requirements.txt
# AlphaOmega Core Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
httpx>=0.25.0

# MCP Server
anthropic>=0.5.0

# Dashboard
flask>=3.0.0
flask-cors>=4.0.0

# Logging
python-json-logger>=2.0.7

# Utilities
pillow>=10.0.0
psutil>=5.9.0
requests>=2.31.0
